+++ packages/terra-core-docs/src/terra-dev-site/doc/content-container/example/ContentContainerFill.jsx
@@ -1,9 +1,9 @@
-import React, { useState } from "react";
-import ContentContainer from "terra-content-container";
-import Button from "terra-button";
-import classNames from "classnames/bind";
-import InputField from "terra-form-input/lib/InputField";
-import styles from "./ContentContainerDocCommon.module.scss";
+import React, { useState } from 'react';
+import ContentContainer from 'terra-content-container';
+import Button from 'terra-button';
+import classNames from 'classnames/bind';
+import InputField from 'terra-form-input/lib/InputField';
+import styles from './ContentContainerDocCommon.module.scss';
 
 const cx = classNames.bind(styles);
 
@@ -22,38 +22,37 @@ const contentheader = <h3>Display Header</h3>;
 const contentfooter = <h3>Display Footer</h3>;
 
 const Container = () => {
-  const [field, setField] = useState("");
+  const [field, setField] = useState('');
 
   const handleFirstChange = (event) => {
     setField(event.target.value);
   };
 
-  const clickHandler = () =>
-    alert(
-      `${
-        field.trim().length !== 0
-          ? `You have entered ${field}`
-          : "You have not entered anything"
-      }`
-    );
+  const clickHandler = () => alert(
+    `${
+      field.trim().length !== 0
+        ? `You have entered ${field}`
+        : 'You have not entered anything'
+    }`,
+  );
 
   return (
-    <div className={cx("content-container-fill")}>
+    <div className={cx('content-container-fill')}>
       <ContentContainer header={contentheader} footer={contentfooter} fill>
         <div key="1">{display1}</div>
         <div key="2">
           <InputField
-            inputId="input-field-2"
-            label="Enter Patient Name"
-            type="text"
+            inputId='input-field-2'
+            label='Enter Patient Name'
+            type='text'
             onChange={handleFirstChange}
           />
         </div>
         <div key="3">
           <Button
-            text="Submit"
+            text='Submit'
             onClick={clickHandler}
-            className={cx("button")}
+            className={cx('button')}
           />
         </div>
       </ContentContainer>