+++ packages/terra-core-docs/src/terra-dev-site/doc/content-container/example/ContentContainerScrollRef.jsx
@@ -1,10 +1,10 @@
-import React, { useState } from "react";
+import React, { useState } from 'react';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
-import Button from "terra-button";
-import ContentContainer from "terra-content-container";
-import InputField from "terra-form-input/lib/InputField";
-import classNames from "classnames/bind";
-import styles from "./ContentContainerDocCommon.module.scss";
+import Button from 'terra-button';
+import ContentContainer from 'terra-content-container';
+import InputField from 'terra-form-input/lib/InputField';
+import classNames from 'classnames/bind';
+import styles from './ContentContainerDocCommon.module.scss';
 
 const cx = classNames.bind(styles);
 
@@ -22,29 +22,27 @@ const display1 = (
 const contentheader = <h3>Display Header</h3>;
 const contentfooter = <h3 id="contentfooter">Children count is </h3>;
 
-const onContainerScroll = (scrollRef) =>
-  document
-    .getElementById("contentfooter")
-    .append(` ${scrollRef.children.length}`);
+const onContainerScroll = (scrollRef) => document
+  .getElementById('contentfooter')
+  .append(` ${scrollRef.children.length}`);
 
 const Container = () => {
-  const [field, setField] = useState("");
+  const [field, setField] = useState('');
 
   const handleFirstChange = (event) => {
     setField(event.target.value);
   };
 
-  const clickHandler = () =>
-    alert(
-      `${
-        field.trim().length !== 0
-          ? `You have entered ${field}`
-          : "You have not entered anything"
-      }`
-    );
+  const clickHandler = () => alert(
+    `${
+      field.trim().length !== 0
+        ? `You have entered ${field}`
+        : 'You have not entered anything'
+    }`,
+  );
 
   return (
-    <div className={cx("content-container-fill")}>
+    <div className={cx('content-container-fill')}>
       {/**
        * ScrollRefCallback is used to refer to the scrollable area of the content container DOM element.
        */}
@@ -57,17 +55,17 @@ const Container = () => {
         <div key="1">{display1}</div>
         <div key="2">
           <InputField
-            inputId="input-field-3"
-            label="Enter Patient Name"
-            type="text"
+            inputId='input-field-3'
+            label='Enter Patient Name'
+            type='text'
             onChange={handleFirstChange}
           />
         </div>
         <div key="3">
           <Button
-            text="Submit"
+            text='Submit'
             onClick={clickHandler}
-            className={cx("button")}
+            className={cx('button')}
           />
         </div>
       </ContentContainer>