+++ packages/terra-core-docs/src/terra-dev-site/doc/content-container/example/ContentContainerStandard.jsx
@@ -1,9 +1,9 @@
-import React, { useState } from "react";
-import ContentContainer from "terra-content-container";
-import classNames from "classnames/bind";
-import InputField from "terra-form-input/lib/InputField";
-import Button from "terra-button";
-import styles from "./ContentContainerDocCommon.module.scss";
+import React, { useState } from 'react';
+import ContentContainer from 'terra-content-container';
+import classNames from 'classnames/bind';
+import InputField from 'terra-form-input/lib/InputField';
+import Button from 'terra-button';
+import styles from './ContentContainerDocCommon.module.scss';
 
 const cx = classNames.bind(styles);
 
@@ -22,26 +22,25 @@ const contentheader = <h3>Display Header</h3>;
 const contentfooter = <h3>Display Footer</h3>;
 
 const Container = () => {
-  const [field, setField] = useState("");
+  const [field, setField] = useState('');
 
   const handleFirstChange = (event) => {
     setField(event.target.value);
   };
 
-  const clickHandler = () =>
-    alert(
-      `${
-        field.trim().length !== 0
-          ? `You have entered ${field}`
-          : "You have not entered anything"
-      }`
-    );
+  const clickHandler = () => alert(
+    `${
+      field.trim().length !== 0
+        ? `You have entered ${field}`
+        : 'You have not entered anything'
+    }`,
+  );
 
   return (
-    <div className={cx("content-container-standard")}>
+    <div className={cx('content-container-standard')}>
       <ContentContainer header={contentheader} footer={contentfooter}>
         <div key="1">{display1}</div>
-        <div key="2" className={cx("field-content")}>
+        <div key="2" className={cx('field-content')}>
           <InputField
             inputId="input-field-1"
             label="Enter Patient Name"